transfer funds by email

+++ db/migrate/20250520145860_create_balance_transactions.rb
@@ -2,7 +2,7 @@ class CreateBalanceTransactions < ActiveRecord::Migration[8.0]
   def change
     create_table :balance_transactions do |t|
       t.references :user, null: false, foreign_key: true
-      t.integer :counterparty_id
+      t.bigint :counterparty_id
       t.integer :amount_cents, null: false
       t.timestamps
     end
